+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
   - Parameters can be now used within conditions.
   - Implementation of `__contains__` to support `if ... in P`.
   - `update()` now takes optional argument `only_existing` (def. `False`) to only update existing parameters.
+  - (ROS1 only) Parameters and their values are exposed to the ROS Parameter Server.
 
 ## 0.5.1 - 2022-03-23
 ### Added